pyproject.toml
[project]
name = "bpma-track-bookings"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4==4.13.5",
    "boto3==1.40.21",
    "google==3.0.0",
    "google-api-python-client==2.179.0",
    "google-genai==1.32.0",
    "openai==1.102.0",
    "pydantic==2.11.7",
    "python-dotenv==1.1.1",
    "requests==2.32.5",
]

[dependency-groups]
dev = [
    "google-api-python-client-stubs==1.30.0",
    "ruff==0.12.11",
    "types-boto3[s3,ssm]==1.40.21",
]
